src/utils/config.py
"""YAML 설정 로더."""

from __future__ import annotations

from pathlib import Path
from typing import Any

import yaml


def load_config(path: str | Path) -> dict[str, Any]:
    """YAML 설정 파일을 읽어 딕셔너리로 반환한다."""
    path = Path(path)
    if not path.exists():
        raise FileNotFoundError(f"설정 파일을 찾을 수 없습니다: {path}")
    with path.open("r", encoding="utf-8") as f:
        data = yaml.safe_load(f)
    if not isinstance(data, dict):
        raise ValueError(f"설정 파일이 올바른 YAML 딕셔너리 형식이 아닙니다: {path}")
    return data
