pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoclip"
version = "0.1.0"
description = "로컬 AI 기반 생방송 자동 하이라이트 클리퍼"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.45.0",
    "accelerate>=0.26.0",
    "faster-whisper>=1.0.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0",
    "tqdm>=4.65.0",
    "huggingface-hub>=0.20.0",
    "numpy>=1.24.0",
    "av>=11.0.0",
    "qwen-vl-utils>=0.0.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
autoclip = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
